--- a/tsfresh/feature_selection/significance_tests.py
+++ b/tsfresh/feature_selection/significance_tests.py
@@ -69,10 +69,10 @@ def target_binary_feature_binary_test(x, y):
     y0, y1 = np.unique(y.values)
 
     # Calculate contingency table
-    n_y1_x0 = np.sum(y.values[x.values == x0] == y1)
-    n_y0_x0 = len(y.values[x.values == x0]) - n_y1_x0
-    n_y1_x1 = np.sum(y.values[x.values == x1] == y1)
-    n_y0_x1 = len(y.values[x.values == x1]) - n_y1_x1
+    n_y1_x0 = np.sum(y[x == x0] == y1)
+    n_y0_x0 = len(y[x == x0]) - n_y1_x0
+    n_y1_x1 = np.sum(y[x == x1] == y1)
+    n_y0_x1 = len(y[x == x1]) - n_y1_x1
 
     table = np.array([[n_y1_x1, n_y1_x0],
                       [n_y0_x1, n_y0_x0]])
@@ -114,8 +114,8 @@ def target_binary_feature_real_test(x, y, test):
     y0, y1 = np.unique(y.values)
 
     # Divide feature according to target
-    x_y1 = x.values[y.values == y1]
-    x_y0 = x.values[y.values == y0]
+    x_y1 = x[y == y1]
+    x_y0 = x[y == y0]
 
     if test == 'mann':
         # Perform Mann-Whitney-U test
@@ -156,8 +156,8 @@ def target_real_feature_binary_test(x, y):
     x0, x1 = np.unique(x.values)
 
     # Divide target according to feature
-    y_x1 = y.values[x.values == x1]
-    y_x0 = y.values[x.values == x0]
+    y_x1 = y[x == x1]
+    y_x0 = y[x == x0]
 
     # Perform Kolmogorov-Smirnov test
     KS, p_value = stats.ks_2samp(y_x1, y_x0)