Changelog:
* Changed all **pandas.Series** methods in the significance tests to **numpy** methods which are faster (see below).

Benchmark example:

```
%timeit x.isnull().any()
573 µs ± 20.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

%timeit numpy.isnan(x.values).any()
15.8 µs ± 418 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```