@MaxBenChrist 2018-12-04
Hey @delyanr, thanks for the great next pr. 

From my experience, speed is not an issue in the tsfresh feature selection, feature extraction is a total different story. Was the feature selection too slow for your application?

Your improvements are sane, but they make the code a little less readable. Also, there could be bugs if x and y are not sorted in the same way. See the following example

```
 import pandas as pd
x = pd.Series([1, 2])
x = x[::-1] # revert x
y = pd.Series([10, 20])
# correct
print(x[y == 10])
# wrong
x.values[y.values == 10]
```
So, I can't merge your pr as it creates possible bugs. 

How can we process from here? You can add an unit test that shows that your implementation has this flaw (so, it should fail). And then afterwards we can fix it by sorting x and y.