@delyanr 2018-12-05
Re speed: I definitely want to improve everything possible, since my dataset is massive and I need to re-run it often (due to changing features, new time series, etc.). Nevertheless feature selection is indeed generally fast, and this will be my last commit to it.

You make a very good point above (somehow I thought sorting will not be an issue). So I suggest we leave the **indexing** based on the current code (using pandas), but still incorporate the **numpy** methods for everything else, e.g. `np.isnan(x.values).any()` instead of `x.isnull().any()`. I believe if we were to sort x and y beforehand that will defeat the speed up.

Let me know if you think that works and I can change the commit.