@MaxBenChrist 2018-12-06
> You make a very good point above (somehow I thought sorting will not be an issue).

I create some nasty bugs in the past because I was so used of pandas aligning the indices automatically when working with np.arrays, so I got sensible to this. 

> Let me know if you think that works and I can change the commit.

sure, we can replace the pandas with the numpy methods